requirements.txt
about-time==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
alive-progress==2.4.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
boto3==1.33.6
botocore==1.33.6
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.0.4
databutton==0.26.11
databutton-web==0.16.0
dataclasses-json==0.5.14
datasets==2.15.0
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.86.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
grapheme==0.6.0
greenlet==3.0.1
h11==0.14.0
htbuilder==0.6.2
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.345
langchain-core==0.0.9
langsmith==0.0.69
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
openai==0.28.1
openai-clip==1.0.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pinecone-client==2.2.4
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==1.10.13
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pypdf==3.17.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.13.2
s3transfer==0.8.2
schedule==1.2.1
sentry-sdk==1.10.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
starlette==0.20.4
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
trio==0.21.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.19.0
validators==0.22.0
watchfiles==0.18.1
watchgod==0.8.2
wcwidth==0.2.12
websockets==10.4
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
